fix for overflows in the renderer when the player goes over a linedef between two sectors of extreme height (see http://www.doomworld.com/vb/showthread.php?s=&threadid=37056) - you still get a big visual glitch, but it's better than a particularly nasty segfault! (adapted from eternity's by RjY)

+++ prboom2/src/r_things.c
@@ -69,8 +69,8 @@ fixed_t pspriteyscale;
 // constant arrays
 //  used for psprite clipping and initializing clipping
 
-short negonearray[MAX_SCREENWIDTH];        // killough 2/8/98:
-short screenheightarray[MAX_SCREENWIDTH];  // change to MAX_*
+int negonearray[MAX_SCREENWIDTH];        // killough 2/8/98: // dropoff overflow
+int screenheightarray[MAX_SCREENWIDTH];  // change to MAX_* // dropoff overflow
 
 //
 // INITIALIZATION FUNCTIONS
@@ -309,8 +309,8 @@ vissprite_t *R_NewVisSprite(void)
 //  in posts/runs of opaque pixels.
 //
 
-short   *mfloorclip;
-short   *mceilingclip;
+int   *mfloorclip;   // dropoff overflow
+int   *mceilingclip; // dropoff overflow
 fixed_t spryscale;
 fixed_t sprtopscreen;
 
@@ -898,8 +898,8 @@ void R_SortVisSprites (void)
 void R_DrawSprite (vissprite_t* spr)
 {
   drawseg_t *ds;
-  short   clipbot[MAX_SCREENWIDTH];       // killough 2/8/98:
-  short   cliptop[MAX_SCREENWIDTH];       // change to MAX_*
+  int     clipbot[MAX_SCREENWIDTH]; // killough 2/8/98: // dropoff overflow
+  int     cliptop[MAX_SCREENWIDTH]; // change to MAX_*  // dropoff overflow
   int     x;
   int     r1;
   int     r2;
